Chatbot/app/requirements.txt
# ── Core API framework ──────────────────────────────────────────
fastapi == 0.135.1
uvicorn == 0.41.0
python-multipart == 0.0.22      

# ── Rate limiting ────────────────────────────────────────────────
slowapi == 0.1.9

# ── Authentication ───────────────────────────────────────────────
python-jose[cryptography] == 3.5.0
passlib[bcrypt] == 1.7.4
bcrypt == 4.0.1

# ── Environment ─────────────────────────────────────────
python-dotenv == 1.1.1

# ── PostgreSQL Database ────────────────────────────────────────
psycopg2-binary == 2.9.11

# ── Vector database ──────────────────────────────────────────────
qdrant-client == 1.17.0

# ── caching & async tasks ───────────────────────────────
redis == 7.2.1
celery == 5.6.2

# ── Embeddings  ──────────────────────────────────────────────
sentence-transformers == 5.2.3
langchain-huggingface == 1.2.1
langchain-qdrant == 1.1.0
langchain-text-splitters == 1.1.1
rank-bm25 == 0.2.2
scikit-learn == 1.8.0
numpy == 2.4.3

# ── Document parsing ───────────────────────────
unstructured[all-docs] == 0.21.5

# ── LLM client ───────────────────────────────────────────────────
groq == 1.0.0

# ── NLP utility ────────────────────────────────────────────────
nltk==3.9.3

# ── For Data validation ──────────────────────────────────────────────
pydantic==2.11.10

# ── Testing Purpose ──────────────────────────────────────────────────────
pytest ==  9.0.2
pytest-asyncio == 1.3.0
httpx == 0.28.1                   